requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
crispy-bootstrap4==2024.1
Django==5.0.3
django-appconf==1.0.6
django-bootstrap4==24.1
django-ckeditor==6.7.1
django-ckeditor-5==0.2.12
django-crispy-forms==2.1
django-dynamic-breadcrumbs==0.4.3
django-extensions==3.2.3
django-fontawesome==1.0
django-fontawesome-5==1.0.18
django-htmx==1.17.3
django-imagefit==0.8.1
django-jazzmin==2.6.0
django-js-asset==2.2.0
django-shell-plus==1.1.7
django-tinymce==3.7.1
django-widget-tweaks==1.5.0
fontawesomefree==6.5.1
humanize==4.9.0
idna==3.6
pillow==10.2.0
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.2.1
